Rename gurdian blade name to snake case

+++ resources/views/components/message.blade.php
@@ -21,7 +21,9 @@
                             <span class="label label-default">Seen</span></strong>
                         @endif
 
-                        <a class="btn btn-danger btn-lg float-right text-white ml-2" type="button" onclick="deleteMsg({{ $message->id }})"><i class="fas fa-trash-alt"></i></a>
+                        @if(current_user()->role == 'student')
+                            <a class="btn btn-danger btn-lg float-right text-white ml-2" type="button" onclick="deleteMsg({{ $message->id }})"><i class="fas fa-trash-alt"></i></a>
+                        @endif
                         @if(!empty($message->file_path) && ($message->file_path) !== null)
                             <a class="btn btn-info btn-lg float-right" href="{{asset($message->file_path)}}" target="_blank">Show Attachment</a>
                         @endif
